split out graphql client into its own module
