+++ tests/unit/conftest.py
@@ -7,9 +7,7 @@ from collections.abc import AsyncIterator, Awaitable, Callable, Iterator
 from functools import partial
 from importlib.metadata import version
 from random import getrandbits
-from typing import Any, Literal, Optional
-from urllib.parse import urljoin
-from uuid import uuid4
+from typing import Any, Literal
 
 import httpx
 import pytest
@@ -19,13 +17,11 @@ from _pytest.terminal import TerminalReporter
 from asgi_lifespan import LifespanManager
 from faker import Faker
 from httpx import AsyncByteStream, Request, Response
-from httpx_ws import AsyncWebSocketSession, aconnect_ws
 from psycopg import Connection
 from pytest_postgresql import factories
 from sqlalchemy import URL, make_url
 from sqlalchemy.ext.asyncio import AsyncEngine, AsyncSession
 from starlette.types import ASGIApp
-from strawberry.subscriptions import GRAPHQL_TRANSPORT_WS_PROTOCOL
 
 import phoenix.trace.v1 as pb
 from phoenix.config import EXPORT_DIR
@@ -41,6 +37,7 @@ from phoenix.server.grpc_server import GrpcServer
 from phoenix.server.types import BatchedCaller, DbSessionFactory
 from phoenix.session.client import Client
 from phoenix.trace.schemas import Span
+from tests.unit.graphql import AsyncGraphQLClient
 from tests.unit.transport import ASGIWebSocketTransport
 
 
@@ -244,7 +241,7 @@ def httpx_client(
 
 
 @pytest.fixture
-def gql_client(httpx_client: httpx.AsyncClient) -> Iterator["AsyncGraphQLClient"]:
+def gql_client(httpx_client: httpx.AsyncClient) -> Iterator[AsyncGraphQLClient]:
     yield AsyncGraphQLClient(httpx_client)
 
 
@@ -351,124 +348,3 @@ def rand_trace_id() -> Iterator[str]:
                 yield span_id
 
     return _(set())
-
-
-class AsyncGraphQLClient:
-    """
-    Async GraphQL client that can execute queries, mutations, and subscriptions.
-    """
-
-    def __init__(
-        self, httpx_client: httpx.AsyncClient, timeout_seconds: Optional[float] = 10
-    ) -> None:
-        self._httpx_client = httpx_client
-        self._timeout_seconds = timeout_seconds
-        self._gql_url = urljoin(str(httpx_client.base_url), "/graphql")
-
-    async def execute(
-        self,
-        query: str,
-        variables: Optional[dict[str, Any]] = None,
-        operation_name: Optional[str] = None,
-    ) -> dict[str, Any]:
-        """
-        Executes queries and mutations.
-        """
-        response = await self._httpx_client.post(
-            self._gql_url,
-            json={
-                "query": query,
-                **({"variables": variables} if variables is not None else {}),
-                **({"operationName": operation_name} if operation_name is not None else {}),
-            },
-        )
-        response.raise_for_status()
-        response_json = response.json()
-        if (errors := response_json.get("errors")) is not None:
-            raise RuntimeError(errors)
-        assert isinstance(data := response_json.get("data"), dict)
-        return data
-
-    @contextlib.asynccontextmanager
-    async def subscription(
-        self,
-        query: str,
-        variables: Optional[dict[str, Any]] = None,
-        operation_name: Optional[str] = None,
-    ) -> AsyncIterator["GraphQLSubscription"]:
-        """
-        Starts a GraphQL subscription session.
-        """
-        async with aconnect_ws(
-            self._gql_url,
-            self._httpx_client,
-            subprotocols=[GRAPHQL_TRANSPORT_WS_PROTOCOL],
-        ) as session:
-            await session.send_json({"type": "connection_init"})
-            message = await session.receive_json(timeout=self._timeout_seconds)
-            if message.get("type") != "connection_ack":
-                raise RuntimeError("Websocket connection failed")
-            yield GraphQLSubscription(
-                session=session,
-                query=query,
-                variables=variables,
-                operation_name=operation_name,
-                timeout_seconds=self._timeout_seconds,
-            )
-
-
-class GraphQLSubscription:
-    """
-    A session for a GraphQL subscription.
-    """
-
-    def __init__(
-        self,
-        *,
-        session: AsyncWebSocketSession,
-        query: str,
-        variables: Optional[dict[str, Any]] = None,
-        operation_name: Optional[str] = None,
-        timeout_seconds: Optional[float] = None,
-    ) -> None:
-        self._session = session
-        self._query = query
-        self._variables = variables
-        self._operation_name = operation_name
-        self._timeout_seconds = timeout_seconds
-
-    async def stream(
-        self,
-    ) -> AsyncIterator[dict[str, Any]]:
-        """
-        Streams subscription payloads.
-        """
-        connection_id = str(uuid4())
-        await self._session.send_json(
-            {
-                "id": connection_id,
-                "type": "subscribe",
-                "payload": {
-                    "query": self._query,
-                    **({"variables": self._variables} if self._variables is not None else {}),
-                    **(
-                        {"operationName": self._operation_name}
-                        if self._operation_name is not None
-                        else {}
-                    ),
-                },
-            }
-        )
-        while True:
-            message = await self._session.receive_json(timeout=self._timeout_seconds)
-            message_type = message.get("type")
-            assert message.get("id") == connection_id
-            if message_type == "complete":
-                break
-            elif message_type == "next":
-                if (data := message["payload"]["data"]) is not None:
-                    yield data
-            elif message_type == "error":
-                raise RuntimeError(message["payload"])
-            else:
-                assert False, f"Unexpected message type: {message_type}"
